Coil added.

+++ app/src/main/java/com/example/cryptoapp/model/home/USD.kt
@@ -11,7 +11,7 @@ data class USD(
     @SerializedName("market_cap")
     val marketCap: Double?,
     @SerializedName("market_cap_dominance")
-    val marketCapDominance: Int?,
+    val marketCapDominance: Double?,
     @SerializedName("percent_change_1h")
     val percentChange1h: Double?,
     @SerializedName("percent_change_24h")
